r/aws_db_proxy: Move functions around.
